Remove unused environments/stages and separate code into routes/modules
