+++ mongo.js
@@ -1,0 +1,13 @@
+'use strict';
+const MongoClient = require('mongodb').MongoClient;
+
+async function connectToMongo() {
+    const client = await MongoClient.connect(process.env.MONGO_URL);
+    return client
+        .db(process.env.MONGO_DB)
+        .collection(process.env.MONGO_COLLECTION);
+}
+
+module.exports = {
+    connectToMongo
+};
